+++ src/components/impact-dashboard/CurveChartLegend.tsx
