+++ src/components/impact-dashboard/EpidemicModelContext.tsx
@@ -2,8 +2,8 @@ import { sum } from "d3";
 import { pick } from "lodash";
 import React from "react";
 
-import { LocaleData } from "../contexts/locale-data-context";
-import { Facility } from "../page-multi-facility/types";
+import { LocaleData } from "../../contexts/locale-data-context";
+import { Facility } from "../../page-multi-facility/types";
 
 export type PlannedRelease = { date?: Date; count?: number };
 export type PlannedReleases = PlannedRelease[];
