+++ src/database/index.tsx
@@ -9,7 +9,7 @@ import { pick, orderBy, uniqBy } from "lodash";
 import { Optional } from "utility-types";
 
 import { MMMMdyyyy } from "../constants";
-import { persistedKeys } from "../impact-dashboard/EpidemicModelContext";
+import { persistedKeys } from "../components/impact-dashboard/EpidemicModelContext";
 import {
   Facility,
   ModelInputs,
