+++ src/page-model-inspection/ModelInspectionTableContainer.tsx
@@ -2,7 +2,7 @@ import { range, sum } from "d3-array";
 import flatten from "lodash/flatten";
 import React from "react";
 
-import { useEpidemicModelState } from "../impact-dashboard/EpidemicModelContext";
+import { useEpidemicModelState } from "../components/impact-dashboard/EpidemicModelContext";
 import {
   calculateAllCurves,
   curveInputsFromUserInputs,
