+++ src/page-multi-facility/FacilityInputForm.tsx
@@ -2,6 +2,9 @@ import { navigate } from "gatsby";
 import React, { useContext, useState } from "react";
 import styled from "styled-components";
 
+import FacilityInformation from "../components/impact-dashboard/FacilityInformation";
+import MitigationInformation from "../components/impact-dashboard/MitigationInformation";
+import useModel from "../components/impact-dashboard/useModel";
 import RtTimeseries from "../components/rt-timeseries";
 import { FacilityContext } from "../contexts/facility-context/FacilityContext";
 import { deleteFacility, saveFacility } from "../database/index";
@@ -15,9 +18,6 @@ import { Column, PageContainer } from "../design-system/PageColumn";
 import PopUpMenu from "../design-system/PopUpMenu";
 import { Spacer } from "../design-system/Spacer";
 import Tooltip from "../design-system/Tooltip";
-import FacilityInformation from "../impact-dashboard/FacilityInformation";
-import MitigationInformation from "../impact-dashboard/MitigationInformation";
-import useModel from "../impact-dashboard/useModel";
 import { updateFacilityRtData } from "../infection-model/rt";
 import AddCasesModal from "./AddCasesModal";
 import FacilityProjections from "./FacilityProjections";
