+++ src/page-multi-facility/FacilityRow.tsx
@@ -3,6 +3,10 @@ import { navigate } from "gatsby";
 import React, { useContext, useState } from "react";
 import styled from "styled-components";
 
+import CurveChartContainer from "../components/impact-dashboard/CurveChartContainer";
+import { totalConfirmedCases } from "../components/impact-dashboard/EpidemicModelContext";
+import { getTotalPopulation } from "../components/impact-dashboard/EpidemicModelContext";
+import useModel from "../components/impact-dashboard/useModel";
 import { FacilityContext } from "../contexts/facility-context/FacilityContext";
 import Colors, { MarkColors as markColors } from "../design-system/Colors";
 import { DateMMMMdyyyy } from "../design-system/DateFormats";
@@ -10,10 +14,6 @@ import FontSizes from "../design-system/FontSizes";
 import iconEditSrc from "../design-system/icons/ic_edit.svg";
 import { Spacer } from "../design-system/Spacer";
 import Tooltip from "../design-system/Tooltip";
-import CurveChartContainer from "../impact-dashboard/CurveChartContainer";
-import { totalConfirmedCases } from "../impact-dashboard/EpidemicModelContext";
-import { getTotalPopulation } from "../impact-dashboard/EpidemicModelContext";
-import useModel from "../impact-dashboard/useModel";
 import { getNewestRt, isRtData } from "../infection-model/rt";
 import AddCasesModal from "./AddCasesModal";
 import FacilityRowRtValuePill from "./FacilityRowRtValuePill";
